refactor(auth): API 키 서비스 통합 및 서비스 계층 권한 검증 추가
