--- src/main/kotlin/team/themoment/datagsm/domain/auth/service/CreateAdminApiKeyService.kt
@@ -1,8 +1,0 @@
-package team.themoment.datagsm.domain.auth.service
-
-import team.themoment.datagsm.domain.auth.dto.request.CreateApiKeyReqDto
-import team.themoment.datagsm.domain.auth.dto.response.ApiKeyResDto
-
-interface CreateAdminApiKeyService {
-    fun execute(reqDto: CreateApiKeyReqDto): ApiKeyResDto
-}
