--- src/main/kotlin/team/themoment/datagsm/domain/auth/service/ModifyAdminApiKeyService.kt
@@ -1,8 +1,0 @@
-package team.themoment.datagsm.domain.auth.service
-
-import team.themoment.datagsm.domain.auth.dto.request.ModifyApiKeyReqDto
-import team.themoment.datagsm.domain.auth.dto.response.ApiKeyResDto
-
-interface ModifyAdminApiKeyService {
-    fun execute(reqDto: ModifyApiKeyReqDto): ApiKeyResDto
-}
