+++ src/test/kotlin/team/themoment/datagsm/domain/auth/service/CreateApiKeyServiceTest.kt
@@ -7,13 +7,18 @@ import io.kotest.matchers.shouldNotBe
 import io.mockk.clearAllMocks
 import io.mockk.every
 import io.mockk.mockk
+import io.mockk.mockkStatic
 import io.mockk.verify
+import org.springframework.security.core.Authentication
+import org.springframework.security.core.context.SecurityContext
+import org.springframework.security.core.context.SecurityContextHolder
 import team.themoment.datagsm.domain.account.entity.AccountJpaEntity
 import team.themoment.datagsm.domain.auth.dto.request.CreateApiKeyReqDto
 import team.themoment.datagsm.domain.auth.entity.ApiKey
 import team.themoment.datagsm.domain.auth.repository.ApiKeyJpaRepository
 import team.themoment.datagsm.domain.auth.service.impl.CreateApiKeyServiceImpl
 import team.themoment.datagsm.global.exception.error.ExpectedException
+import team.themoment.datagsm.global.security.checker.ScopeChecker
 import team.themoment.datagsm.global.security.data.ApiKeyEnvironment
 import team.themoment.datagsm.global.security.provider.CurrentUserProvider
 import java.time.LocalDateTime
@@ -26,12 +31,14 @@ class CreateApiKeyServiceTest :
         val mockApiKeyRepository = mockk<ApiKeyJpaRepository>()
         val mockCurrentUserProvider = mockk<CurrentUserProvider>()
         val mockApiKeyEnvironment = mockk<ApiKeyEnvironment>()
+        val mockScopeChecker = mockk<ScopeChecker>()
 
         val createApiKeyService =
             CreateApiKeyServiceImpl(
                 mockApiKeyRepository,
                 mockCurrentUserProvider,
                 mockApiKeyEnvironment,
+                mockScopeChecker,
             )
 
         afterEach {
@@ -46,10 +53,16 @@ class CreateApiKeyServiceTest :
                         id = 1L
                         email = "test@gsm.hs.kr"
                     }
+                val mockAuthentication = mockk<Authentication>()
+                val mockSecurityContext = mockk<SecurityContext>()
 
                 beforeEach {
+                    mockkStatic(SecurityContextHolder::class)
+                    every { SecurityContextHolder.getContext() } returns mockSecurityContext
+                    every { mockSecurityContext.authentication } returns mockAuthentication
                     every { mockCurrentUserProvider.getCurrentAccount() } returns mockAccount
                     every { mockApiKeyEnvironment.expirationDays } returns 30L
+                    every { mockScopeChecker.hasScope(mockAuthentication, "admin:apikey") } returns false
                 }
 
                 context("기존 API 키가 없을 때") {
@@ -244,6 +257,75 @@ class CreateApiKeyServiceTest :
 
                         exception.statusCode.value() shouldBe 400
                         exception.message shouldBe "일반 사용자는 READ scope만 사용 가능합니다. 사용 불가능한 scope: club:write"
+
+                        verify(exactly = 0) { mockApiKeyRepository.save(any()) }
+                    }
+                }
+
+                context("Admin 사용자가 모든 scope로 API 키를 생성할 때") {
+                    val reqDto =
+                        CreateApiKeyReqDto(
+                            scopes =
+                                setOf(
+                                    "student:read",
+                                    "student:write",
+                                    "club:read",
+                                    "club:write",
+                                    "project:read",
+                                    "project:write",
+                                    "admin:excel",
+                                ),
+                            description = "Admin API 키",
+                        )
+
+                    beforeEach {
+                        every { mockScopeChecker.hasScope(mockAuthentication, "admin:apikey") } returns true
+                        every { mockApiKeyEnvironment.adminExpirationDays } returns 365L
+                        every { mockApiKeyRepository.findByAccount(mockAccount) } returns Optional.empty()
+                        every { mockApiKeyRepository.save(any()) } answers {
+                            val entity = firstArg<ApiKey>()
+                            entity.apply { id = 1L }
+                        }
+                    }
+
+                    it("모든 scope로 API 키를 생성하고 365일 만료기간을 설정해야 한다") {
+                        val beforeExecution = LocalDateTime.now()
+                        val result = createApiKeyService.execute(reqDto)
+                        val afterExecution = LocalDateTime.now()
+
+                        result.apiKey shouldNotBe null
+                        result.scopes shouldBe reqDto.scopes
+                        result.description shouldBe reqDto.description
+
+                        val expectedMinExpiresAt = beforeExecution.plusDays(365)
+                        val expectedMaxExpiresAt = afterExecution.plusDays(365)
+                        result.expiresAt.isAfter(expectedMinExpiresAt.minusSeconds(1)) shouldBe true
+                        result.expiresAt.isBefore(expectedMaxExpiresAt.plusSeconds(1)) shouldBe true
+
+                        verify(exactly = 1) { mockApiKeyRepository.save(any()) }
+                    }
+                }
+
+                context("Admin 사용자가 유효하지 않은 scope를 요청할 때") {
+                    val reqDto =
+                        CreateApiKeyReqDto(
+                            scopes = setOf("invalid:scope"),
+                            description = "테스트",
+                        )
+
+                    beforeEach {
+                        every { mockScopeChecker.hasScope(mockAuthentication, "admin:apikey") } returns true
+                        every { mockApiKeyRepository.findByAccount(mockAccount) } returns Optional.empty()
+                    }
+
+                    it("400 BAD_REQUEST 예외가 발생해야 한다") {
+                        val exception =
+                            shouldThrow<ExpectedException> {
+                                createApiKeyService.execute(reqDto)
+                            }
+
+                        exception.statusCode.value() shouldBe 400
+                        exception.message shouldBe "유효하지 않은 scope입니다: invalid:scope"
 
                         verify(exactly = 0) { mockApiKeyRepository.save(any()) }
                     }
