+++ src/test/kotlin/team/themoment/datagsm/domain/auth/service/ModifyApiKeyServiceTest.kt
@@ -7,13 +7,18 @@ import io.kotest.matchers.shouldNotBe
 import io.mockk.clearAllMocks
 import io.mockk.every
 import io.mockk.mockk
+import io.mockk.mockkStatic
 import io.mockk.verify
+import org.springframework.security.core.Authentication
+import org.springframework.security.core.context.SecurityContext
+import org.springframework.security.core.context.SecurityContextHolder
 import team.themoment.datagsm.domain.account.entity.AccountJpaEntity
 import team.themoment.datagsm.domain.auth.dto.request.ModifyApiKeyReqDto
 import team.themoment.datagsm.domain.auth.entity.ApiKey
 import team.themoment.datagsm.domain.auth.repository.ApiKeyJpaRepository
 import team.themoment.datagsm.domain.auth.service.impl.ModifyApiKeyServiceImpl
 import team.themoment.datagsm.global.exception.error.ExpectedException
+import team.themoment.datagsm.global.security.checker.ScopeChecker
 import team.themoment.datagsm.global.security.data.ApiKeyEnvironment
 import team.themoment.datagsm.global.security.provider.CurrentUserProvider
 import java.time.LocalDateTime
@@ -26,12 +31,14 @@ class ModifyApiKeyServiceTest :
         val mockApiKeyRepository = mockk<ApiKeyJpaRepository>()
         val mockCurrentUserProvider = mockk<CurrentUserProvider>()
         val mockApiKeyEnvironment = mockk<ApiKeyEnvironment>()
+        val mockScopeChecker = mockk<ScopeChecker>()
 
         val modifyApiKeyService =
             ModifyApiKeyServiceImpl(
                 mockApiKeyRepository,
                 mockCurrentUserProvider,
                 mockApiKeyEnvironment,
+                mockScopeChecker,
             )
 
         afterEach {
@@ -46,11 +53,17 @@ class ModifyApiKeyServiceTest :
                         id = 1L
                         email = "test@gsm.hs.kr"
                     }
+                val mockAuthentication = mockk<Authentication>()
+                val mockSecurityContext = mockk<SecurityContext>()
 
                 beforeEach {
+                    mockkStatic(SecurityContextHolder::class)
+                    every { SecurityContextHolder.getContext() } returns mockSecurityContext
+                    every { mockSecurityContext.authentication } returns mockAuthentication
                     every { mockCurrentUserProvider.getCurrentAccount() } returns mockAccount
                     every { mockApiKeyEnvironment.expirationDays } returns 30L
                     every { mockApiKeyEnvironment.renewalPeriodDays } returns 15L
+                    every { mockScopeChecker.hasScope(mockAuthentication, "admin:apikey") } returns false
                 }
 
                 context("API 키를 찾을 수 없을 때") {
@@ -365,6 +378,137 @@ class ModifyApiKeyServiceTest :
 
                         exception.statusCode.value() shouldBe 400
                         exception.message shouldBe "일반 사용자는 READ scope만 사용 가능합니다. 사용 불가능한 scope: project:*"
+
+                        verify(exactly = 0) { mockApiKeyRepository.save(any()) }
+                        verify(exactly = 0) { mockApiKeyRepository.delete(any()) }
+                    }
+                }
+
+                context("Admin 사용자가 모든 scope로 갱신할 때") {
+                    val reqDto =
+                        ModifyApiKeyReqDto(
+                            scopes =
+                                setOf(
+                                    "student:read",
+                                    "student:write",
+                                    "club:read",
+                                    "club:write",
+                                    "project:read",
+                                    "project:write",
+                                    "admin:excel",
+                                ),
+                            description = "Admin API 키 갱신",
+                        )
+                    val now = LocalDateTime.now()
+                    val expiresAt = now.plusDays(20)
+                    val oldApiKeyValue = UUID.randomUUID()
+                    val apiKey =
+                        ApiKey().apply {
+                            id = 1L
+                            value = oldApiKeyValue
+                            account = mockAccount
+                            createdAt = now.minusDays(345)
+                            updatedAt = now.minusDays(345)
+                            this.expiresAt = expiresAt
+                        }
+
+                    beforeEach {
+                        every { mockScopeChecker.hasScope(mockAuthentication, "admin:apikey") } returns true
+                        every { mockApiKeyEnvironment.adminExpirationDays } returns 365L
+                        every { mockApiKeyEnvironment.adminRenewalPeriodDays } returns 30L
+                        every { mockApiKeyRepository.findByAccount(mockAccount) } returns Optional.of(apiKey)
+                        every { mockApiKeyRepository.save(apiKey) } returns apiKey
+                    }
+
+                    it("모든 scope로 갱신하고 365일 만료기간을 설정해야 한다") {
+                        val beforeExecution = LocalDateTime.now()
+                        val result = modifyApiKeyService.execute(reqDto)
+                        val afterExecution = LocalDateTime.now()
+
+                        result.apiKey shouldBe oldApiKeyValue
+                        result.scopes shouldBe reqDto.scopes
+                        result.description shouldBe reqDto.description
+
+                        val expectedMinExpiresAt = beforeExecution.plusDays(365)
+                        val expectedMaxExpiresAt = afterExecution.plusDays(365)
+                        result.expiresAt.isAfter(expectedMinExpiresAt.minusSeconds(1)) shouldBe true
+                        result.expiresAt.isBefore(expectedMaxExpiresAt.plusSeconds(1)) shouldBe true
+
+                        verify(exactly = 1) { mockApiKeyRepository.save(apiKey) }
+                        verify(exactly = 0) { mockApiKeyRepository.delete(any()) }
+                    }
+                }
+
+                context("Admin 사용자가 갱신 기간 외에 갱신을 시도할 때 (만료 40일 전)") {
+                    val reqDto =
+                        ModifyApiKeyReqDto(
+                            scopes = setOf("student:read"),
+                            description = "테스트",
+                        )
+                    val now = LocalDateTime.now()
+                    val expiresAt = now.plusDays(40)
+                    val apiKey =
+                        ApiKey().apply {
+                            id = 1L
+                            value = UUID.randomUUID()
+                            account = mockAccount
+                            createdAt = now.minusDays(325)
+                            updatedAt = now.minusDays(325)
+                            this.expiresAt = expiresAt
+                        }
+
+                    beforeEach {
+                        every { mockScopeChecker.hasScope(mockAuthentication, "admin:apikey") } returns true
+                        every { mockApiKeyEnvironment.adminRenewalPeriodDays } returns 30L
+                        every { mockApiKeyRepository.findByAccount(mockAccount) } returns Optional.of(apiKey)
+                    }
+
+                    it("400 BAD_REQUEST 예외가 발생해야 한다") {
+                        val exception =
+                            shouldThrow<ExpectedException> {
+                                modifyApiKeyService.execute(reqDto)
+                            }
+
+                        exception.statusCode.value() shouldBe 400
+                        exception.message shouldBe "API 키 갱신 기간이 아닙니다. 만료 30일 전부터 만료 30일 후까지만 갱신 가능합니다."
+
+                        verify(exactly = 0) { mockApiKeyRepository.save(any()) }
+                        verify(exactly = 0) { mockApiKeyRepository.delete(any()) }
+                    }
+                }
+
+                context("Admin 사용자가 유효하지 않은 scope로 갱신을 시도할 때") {
+                    val reqDto =
+                        ModifyApiKeyReqDto(
+                            scopes = setOf("invalid:scope"),
+                            description = "테스트",
+                        )
+                    val now = LocalDateTime.now()
+                    val expiresAt = now.plusDays(10)
+                    val apiKey =
+                        ApiKey().apply {
+                            id = 1L
+                            value = UUID.randomUUID()
+                            account = mockAccount
+                            createdAt = now.minusDays(355)
+                            updatedAt = now.minusDays(355)
+                            this.expiresAt = expiresAt
+                        }
+
+                    beforeEach {
+                        every { mockScopeChecker.hasScope(mockAuthentication, "admin:apikey") } returns true
+                        every { mockApiKeyEnvironment.adminRenewalPeriodDays } returns 30L
+                        every { mockApiKeyRepository.findByAccount(mockAccount) } returns Optional.of(apiKey)
+                    }
+
+                    it("400 BAD_REQUEST 예외가 발생해야 한다") {
+                        val exception =
+                            shouldThrow<ExpectedException> {
+                                modifyApiKeyService.execute(reqDto)
+                            }
+
+                        exception.statusCode.value() shouldBe 400
+                        exception.message shouldBe "유효하지 않은 scope입니다: invalid:scope"
 
                         verify(exactly = 0) { mockApiKeyRepository.save(any()) }
                         verify(exactly = 0) { mockApiKeyRepository.delete(any()) }
